pyproject.toml
[project]
name = "scatterlabel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.3.0",
    "psycopg>=3.2.9",
    "python-dotenv>=1.1.0",
    "tqdm>=4.67.1",
    "requests>=2.32.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.5.0",
]

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "pandas-stubs>=2.2.3.250527",
    "types-psycopg2>=2.9.21.20250516",
    "types-tqdm>=4.67.0.20250516",
]
